requirements/dev.txt
-r base.txt

django-autofixture==0.6.1
django-reset==0.1.0
pudb==2013.3.6
coverage==3.6
django-discover-runner==1.0
django-debug-toolbar==0.9.4
django-extensions==1.2.5
dj-database-url
django-toolbelt